eastwind: the declarations Southwind makes and eastwind had not
Six `[UniqueIndex]` attributes Southwind puts on the natural key of its master
data — Region and Territory descriptions, Category, Product, Supplier and Shipper
names — and the two mixins its Starter registers:

  • `CorruptMixin` on PersonEntity. The person is the app's one example of a row
    that may be saved INVALID (an imported legacy customer with no title or birth
    date), which is what the mixin's `corrupt` flag records. Southwind's
    `IsApplicableValidator(p => Corruption.Strict)` on Title / DateOfBirth is the
    other half; altea has no Corruption scope, so the column is carried and the
    escape hatch is not.

  • `OrderDetailMixin` on the order line, kept under SOUTHWIND's name because a
    mixin's name is identity in a property route — `[OrderDetailMixin].discountCode`
    is what a property-auth rule and a translated instance store — where the
    entity it hangs off was free to be renamed (OrderDetailEmbedded ->
    OrderLineEntity), since altea's part rows carry their own table.

PersonEntity's `toString()` loses its `@quoted`: Southwind writes a plain
`ToString()` there, not an [AutoExpressionField], so the display string is STORED
in a `to_str` column rather than expanded into every query. CompanyEntity right
below it is the other case and keeps its `@quoted`.

`ProductPredictorPublication.MonthlySales` is declared where Southwind declares
it and registered in the starter, which is also what seeds its symbol row.
`Product.unitsInStock` is a `short`, as Northwind's column is, and the loader and
the raw-schema row follow. `ApplicationConfiguration.emailSender` is non-null.
`OrderState.New` is marked not-mapped — Southwind's [Ignore] — so it is not a row
of the enum table shifting every following id by one.

+++ eastwind/terminal/northwindSchema.ts
@@ -1,7 +1,7 @@
 import { reflect } from "@altea/altea/data/reflection";
 import { tableName, viewPrimaryKey } from "@altea/altea/data/decorators";
 import { View } from "@altea/altea/data/entity";
-import { Temporal, type int, Decimal } from "@altea/altea/data/basics";
+import { Temporal, type int, type short, Decimal } from "@altea/altea/data/basics";
 import { SchemaBuilder } from "@altea/altea/server/schema";
 import { ViewBuilder } from "@altea/altea/server/schema/viewBuilder";
 import { ObjectName, SchemaName, DatabaseName } from "@altea/altea/server/schema/objectName";
@@ -93,7 +93,7 @@ export class NwProduct extends View {
     CategoryID!: int | null;
     QuantityPerUnit!: string | null;
     UnitPrice!: Decimal | null;
-    UnitsInStock!: int | null;
+    UnitsInStock!: short | null;
     ReorderLevel!: int | null;
     Discontinued!: boolean;
 }
